+++ packages/core/utils/src/common/plurailze.ts
@@ -1,6 +1,5 @@
 import pluralizeEN from "pluralize"
 pluralizeEN.addUncountableRule("info")
-
 /**
  * Function to pluralize English words.
  * @param word
@@ -8,4 +7,8 @@ pluralizeEN.addUncountableRule("info")
 export function pluralize(word: string): string {
   // TODO: Implement language specific pluralize function
   return pluralizeEN(word)
+}
+
+export function singularize(word: string): string {
+  return pluralizeEN.singular(word)
 }